soc/intel: Adapt crashlog IP to also support 64-bit
This patch extends the crashlog IP support beyond 32-bit mode to
support Intel future generation SoCs, which may require crashlog
support for 64-bit architectures. uintptr_t data type is used for
Address pointers and void* for dereferencing

BUG=b:346676856
TEST=Successfully built Meteor Lake (rex) and tested for google/rex0
and google/rex64 images.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/83106
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/soc/intel/common/block/include/intelblocks/crashlog.h
@@ -181,9 +181,9 @@ void free_cl_node(cl_node_t *node);
 int cl_get_cpu_record_size(void);
 int cl_get_pmc_record_size(void);
 int cl_get_ioe_record_size(void);
-u32 cl_get_cpu_bar_addr(void);
-u32 cl_get_cpu_tmp_bar(void);
-u32 cl_get_cpu_mb_int_addr(void);
+uintptr_t cl_get_cpu_bar_addr(void);
+uintptr_t cl_get_cpu_tmp_bar(void);
+uintptr_t cl_get_cpu_mb_int_addr(void);
 int cl_get_total_data_size(void);
 bool cl_pmc_sram_has_mmio_access(void);
 bool cl_ioe_sram_has_mmio_access(void);
@@ -201,13 +201,13 @@ void update_new_ioe_crashlog_size(u32 *pmc_crash_size);
 pmc_ipc_discovery_buf_t cl_get_pmc_discovery_buf(void);
 pmc_crashlog_desc_table_t cl_get_pmc_descriptor_table(void);
 cpu_crashlog_discovery_table_t cl_get_cpu_discovery_table(void);
-u32 cl_gen_cpu_bar_addr(void);
-int cpu_cl_poll_mailbox_ready(u32 cl_mailbox_addr);
+uintptr_t cl_gen_cpu_bar_addr(void);
+int cpu_cl_poll_mailbox_ready(uintptr_t cl_mailbox_addr);
 int cpu_cl_mailbox_cmd(u8 cmd, u8 param);
 int cpu_cl_clear_data(void);
 void cpu_cl_rearm(void);
 void cpu_cl_cleanup(void);
-int pmc_cl_gen_descriptor_table(u32 desc_table_addr,
+int pmc_cl_gen_descriptor_table(uintptr_t desc_table_addr,
 				pmc_crashlog_desc_table_t *descriptor_table);
 bool pmc_cl_discovery(void);
 bool cpu_cl_discovery(void);
@@ -215,12 +215,8 @@ int cl_pmc_re_arm_after_reset(void);
 int cl_pmc_clear(void);
 int cl_pmc_en_gen_on_all_reboot(void);
 bool discover_crashlog(void);
-bool cl_copy_data_from_sram(u32 src_bar,
-			u32 offset,
-			u32 size,
-			u32 *dest_addr,
-			u32 buffer_index,
-			bool pmc_sram);
+bool cl_copy_data_from_sram(uintptr_t src_bar, u32 offset, size_t size, u32 *dest_addr,
+				u32 buffer_index, bool pmc_sram);
 void collect_pmc_and_cpu_crashlog_from_srams(cl_node_t *head);
 static const EFI_GUID FW_ERR_SECTION_GUID = {
 	0x81212a96, 0x09ed, 0x4996,
